Rain support for iRacing Part 1

+++ Sources/Special/IRC SHM Spotter/IRC SHM Spotter/irsdk_utils.cpp
@@ -338,7 +338,7 @@ unsigned int irsdk_getBroadcastMsgID()
 
 void irsdk_broadcastMsg(irsdk_BroadcastMsg msg, int var1, int var2, int var3)
 {
-	irsdk_broadcastMsg(msg, var1, MAKELONG(var2, var3));
+	irsdk_broadcastMsg(msg, var1, (int)MAKELONG(var2, var3));
 }
 
 void irsdk_broadcastMsg(irsdk_BroadcastMsg msg, int var1, float var2)
